refactor: move coordinated leadership transfer code into its own package
Moves the bulk of the leader-side CLT code from `io.atomix.raft.role`
into `io.atomix.raft.rebalance`, now that it's no longer so closely tied
into `LeaderRole`.

`LeadershipTransferRunner` is the entry point for `LeaderRole` (and
`LeadershipTransferPauseGuard`, which stays owned by the role) to call
in to the new package.
