+++ zeebe/atomix/cluster/src/main/java/io/atomix/raft/rebalance/LeadershipTransferAttempt.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.raft.roles;
+package io.atomix.raft.rebalance;
 
 import com.google.common.base.Throwables;
 import io.atomix.cluster.MemberId;
@@ -21,6 +21,7 @@ import io.atomix.raft.LeadershipTransferResult;
 import io.atomix.raft.impl.RaftContext;
 import io.atomix.raft.protocol.LeadershipTransferInitiateRequest;
 import io.atomix.raft.protocol.LeadershipTransferResultRequest;
+import io.atomix.raft.roles.LeaderRole;
 import java.time.Duration;
 import java.util.concurrent.CompletableFuture;
 import org.jspecify.annotations.NullMarked;
