+++ zeebe/atomix/cluster/src/test/java/io/atomix/raft/rebalance/CatchUpWaitTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.atomix.raft.roles;
+package io.atomix.raft.rebalance;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -21,6 +21,7 @@ import io.atomix.cluster.MemberId;
 import io.atomix.raft.LeadershipTransferResult;
 import io.atomix.raft.RaftRule;
 import io.atomix.raft.RaftServer;
+import io.atomix.raft.roles.LeaderRole;
 import java.time.Duration;
 import java.util.Optional;
 import java.util.concurrent.CompletableFuture;
